@@ -13,14 +13,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.4.0",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@sentry/angular-ivy": "^7.61.0",
     "drift-zoom": "^1.4.0",
     "exifr": "^7.1.3",
@@ -35,10 +35,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/estree": "^0.0.51",
     "@types/google.accounts": "^0.0.2",
     "@types/jasmine": "~3.6.0",
@@ -55,7 +55,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.5.0",
-    "ng-packagr": "~13.3.1",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
